Working on multiple subprocesses.
Things such as pipewire need to be executed through the window manager.
Not to mention somebar, a status bar application for this window manager
will be executed through there. Then a foot terminal server will be
executed through here as well. All in all 3 processes, but one will be
reserved. Eventually somebar will be forked and rewritten in C because
there are no C dependencies.